html/dom/documents/resource-metadata-management/document-readyState.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>document.readyState</title>
<link rel="author" title="Denis Ah-Kang" href="mailto:denis@w3.org">
<link rel=help href="http://www.w3.org/html/wg/drafts/html/master/#resource-metadata-management">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
  var t = async_test("readyState equals 'complete' when the document has loaded");
  var t2 = async_test("readystatechange event is fired each time document.readyState changes");

  window.onload = t.step_func_done(function(){
    assert_equals(document.readyState, "complete");
  });

  var states = ["loading"],
      loop = true;
  document.onreadystatechange = t2.step_func(function(){
    states.push(document.readyState);
    if (document.readyState === "complete") {
      assert_array_equals(states, ["loading", "interactive", "complete"]);
      t2.done();
    }
  })
</script>
